Add refreshing res_state method
By setting the parameter '''refreshing''' to True, the reservoir will iterate the reservoir state after each one-step prediction. With this method, we train the weight matrix by setting prediction horizon to 1, and we can predict any length by one train. So the parameter '''n_test''' determines how many steps it predicts. Set the '''refreshing''' to False then it should be the same as the old version. The parameters still need to be optimized.