experiment scripts
This new branch is for the experiment scripts. The main difference is that we do the iteration in matlab, to improve the efficiency. Also a few more functions are added to 'iterate', which is controled by the variable 'iterate_type'. The details are contained in the matlab code as comments.